Lxpscrum 804 configurer les endpoints api pour que un etudiant puisse creer des tags

+++ api/src/models/user/delete-user.ts
@@ -1,11 +1,17 @@
 import { prisma } from "../../utils/db";
 import User from "../../utils/interfaces/db/user";
 
-export default async function deleteUser(userId: string) {
-  await prisma.student.deleteMany({ where: { idMdb: userId } });
-  await prisma.admin.deleteMany({ where: { idMdb: userId } });
+export default async function deleteUser(userId: string, connectedId: string) {
+  console.log({ userId, connectedId });
 
-  await User.deleteOne().where({ _id: userId });
+  if (userId === connectedId)
+    throw new Error("Vous ne pouvez pas vous supprimer vous même");
 
-  return;
+  try {
+    await prisma.student.deleteMany({ where: { idMdb: userId } });
+    await prisma.admin.deleteMany({ where: { idMdb: userId } });
+    await User.deleteOne().where({ _id: userId });
+  } catch (error) {
+    throw new Error("Cet utilisateur ne peut pas être supprimé");
+  }
 }
